requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiohttp==3.9.3
    # via
    #   datasets
    #   fsspec
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via langchain-core
astroid==3.1.0
    # via pylint
asttokens==2.4.1
    # via stack-data
attrs==23.2.0
    # via aiohttp
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via nltk
comm==0.2.2
    # via ipywidgets
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dataclasses-json==0.6.4
    # via
    #   langchain
    #   langchain-community
datasets==2.14.4
    # via evaluate
decorator==5.1.1
    # via ipython
deepl==1.17.0
    # via lingo (pyproject.toml)
dill==0.3.7
    # via
    #   datasets
    #   evaluate
    #   multiprocess
    #   pylint
evaluate==0.4.1
    # via lingo (pyproject.toml)
executing==2.0.1
    # via stack-data
filelock==3.13.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.51.0
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.3.1
    # via
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   torch
google-api-core[grpc]==2.18.0
    # via
    #   google-cloud-core
    #   google-cloud-translate
google-auth==2.29.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-translate
google-cloud-core==2.4.1
    # via google-cloud-translate
google-cloud-translate==3.15.3
    # via lingo (pyproject.toml)
googleapis-common-protos==1.63.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.62.2
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.2
    # via google-api-core
huggingface-hub==0.22.2
    # via
    #   datasets
    #   evaluate
    #   tokenizers
    #   transformers
idna==3.6
    # via
    #   anyio
    #   requests
    #   yarl
ipython==8.23.0
    # via ipywidgets
ipywidgets==8.1.2
    # via lingo (pyproject.toml)
isort==5.13.2
    # via pylint
jedi==0.19.1
    # via ipython
jinja2==3.1.3
    # via torch
joblib==1.4.0
    # via nltk
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
jupyterlab-widgets==3.0.10
    # via ipywidgets
kiwisolver==1.4.5
    # via matplotlib
langchain==0.1.12
    # via lingo (pyproject.toml)
langchain-community==0.0.28
    # via langchain
langchain-core==0.1.32
    # via
    #   langchain
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.0.1
    # via langchain
langsmith==0.1.28
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markupsafe==2.1.5
    # via jinja2
marshmallow==3.21.1
    # via dataclasses-json
matplotlib==3.8.4
    # via lingo (pyproject.toml)
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via pylint
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.15
    # via
    #   datasets
    #   evaluate
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.3
    # via torch
nltk==3.8.1
    # via lingo (pyproject.toml)
numpy==1.26.4
    # via
    #   contourpy
    #   datasets
    #   evaluate
    #   langchain
    #   langchain-community
    #   matplotlib
    #   pandas
    #   pyarrow
    #   torchvision
    #   transformers
orjson==3.9.15
    # via langsmith
packaging==23.2
    # via
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
    #   matplotlib
    #   transformers
pandas==2.2.2
    # via
    #   datasets
    #   evaluate
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.3.0
    # via
    #   matplotlib
    #   torchvision
platformdirs==4.2.1
    # via pylint
prompt-toolkit==3.0.43
    # via ipython
proto-plus==1.23.0
    # via
    #   google-api-core
    #   google-cloud-translate
protobuf==4.25.3
    # via
    #   google-api-core
    #   google-cloud-translate
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyarrow==15.0.2
    # via datasets
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.6.4
    # via
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.16.3
    # via pydantic
pygments==2.17.2
    # via ipython
pylint==3.1.0
    # via lingo (pyproject.toml)
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via lingo (pyproject.toml)
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   datasets
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   transformers
regex==2023.12.25
    # via
    #   nltk
    #   transformers
requests==2.31.0
    # via
    #   datasets
    #   deepl
    #   evaluate
    #   google-api-core
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
    #   responses
    #   transformers
responses==0.18.0
    # via evaluate
rouge==1.0.1
    # via lingo (pyproject.toml)
rsa==4.9
    # via google-auth
safetensors==0.4.2
    # via transformers
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
    #   rouge
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.28
    # via
    #   langchain
    #   langchain-community
stack-data==0.6.3
    # via ipython
sympy==1.12
    # via torch
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tokenizers==0.15.2
    # via transformers
tomlkit==0.12.4
    # via pylint
torch==2.2.2
    # via
    #   lingo (pyproject.toml)
    #   torchaudio
    #   torchvision
torchaudio==2.2.2
    # via lingo (pyproject.toml)
torchvision==0.17.2
    # via lingo (pyproject.toml)
tqdm==4.66.2
    # via
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   nltk
    #   transformers
traitlets==5.14.2
    # via
    #   comm
    #   ipython
    #   ipywidgets
    #   matplotlib-inline
transformers==4.39.3
    # via lingo (pyproject.toml)
typing-extensions==4.10.0
    # via
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via
    #   requests
    #   responses
wcwidth==0.2.13
    # via prompt-toolkit
widgetsnbextension==4.0.10
    # via ipywidgets
xxhash==3.4.1
    # via
    #   datasets
    #   evaluate
yarl==1.9.4
    # via aiohttp
